refactor(frontend): split DriveExplorer into hooks and views
DriveExplorer.tsx: 2587 -> 409 lines, across 26 flat siblings (the Drives
folder has no subdirectories; new modules join it as siblings).

  views    DriveBreadcrumb, DriveTreeRow, DriveFileContentViewer,
           DriveFilePreview, FolderTile, FolderView, DriveHeader,
           DriveTreeList, DriveTreePane, DriveToolbar, DriveExplorerStates,
           DrivePendingTiles
  hooks    useDriveFilters, useDriveSelection, useDriveTreeData,
           useDriveTreePane, useTreeGroupScroll, useDriveTreeViewport,
           useDriveTreeReveal, useDriveTreeKeyboard, useDriveDownloadAll,
           useDriveUploads
  helpers  driveTypes, driveTreeView, useDelayedTrue; motion tokens append
           to the existing driveMotion.ts

Refs stay inside their owning hook: the scroll element never leaves
useDriveTreeViewport (group scroll attaches via attachTreeWheel) and the
group maps never leave useTreeGroupScroll (rows read scrollXFor(parent)).

Split against current content so the upload/staging feature survives.
useDriveUploads is called ABOVE the tree so its uploadFiles feed
buildDriveTree — that is what makes an in-flight upload render as a real
row under its destination folder rather than a pinned list — and
pendingUploadByPath then decorates the resulting rows. retry/dismiss are
exposed as named accessors rather than leaking the upload object.

currentFolder and commitStaged stay in the root: they read the selection,
which derives from the tree, which consumes useDriveUploads — moving them
into the hook would be circular. Verified line-by-line against the upstream
diff.
