+++ web/oss/src/components/Drives/useDriveDownloadAll.ts
@@ -1,0 +1,51 @@
+/**
+ * useDriveDownloadAll — the header's "Download all": which mounts the ONE zip spans, the in-flight
+ * flag, and the toast-reporting handler (including the failure path's message).
+ */
+import {useCallback, useMemo, useState} from "react"
+
+import {App} from "antd"
+
+import {driveRootLabel} from "./DriveBreadcrumb"
+import {downloadMountArchive} from "./driveMedia"
+import {AGENT_FILES_DIR, type SessionDriveData} from "./useSessionDrive"
+
+export function useDriveDownloadAll({
+    drive,
+    projectId,
+}: {
+    drive: SessionDriveData
+    projectId: string | null | undefined
+}) {
+    const {message} = App.useApp()
+    // "Download all" — ONE streaming zip spanning every mount the drive folds in (cwd at the root,
+    // the agent's durable folder under `agent-files/`). The toast rides App.useApp (dark-mode safe).
+    const [downloadingAll, setDownloadingAll] = useState(false)
+    const archiveMounts = useMemo(() => {
+        const cwd = drive.mount
+        if (!cwd?.id) return []
+        const out = [{mountId: cwd.id, prefix: ""}]
+        const agent = drive.resolveMount(AGENT_FILES_DIR)
+        if (agent && agent.mount.id !== cwd.id) {
+            out.push({mountId: agent.mount.id, prefix: AGENT_FILES_DIR})
+        }
+        return out
+    }, [drive])
+    const handleDownloadAll = useCallback(async () => {
+        if (!archiveMounts.length || downloadingAll) return
+        setDownloadingAll(true)
+        const key = "drive-download-all"
+        message.open({type: "loading", key, content: "Preparing download…", duration: 0})
+        const result = await downloadMountArchive({
+            mounts: archiveMounts,
+            projectId,
+            filename: `${driveRootLabel(drive.mount)}-files.zip`,
+        })
+        if (result.cancelled) message.destroy(key)
+        else if (result.ok) message.open({type: "success", key, content: "Download ready"})
+        else message.open({type: "error", key, content: result.error ?? "Download failed"})
+        setDownloadingAll(false)
+    }, [archiveMounts, drive.mount, projectId, downloadingAll, message])
+
+    return {archiveMounts, downloadingAll, handleDownloadAll}
+}
